+++ bin/eks-cdk.ts
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
 import 'source-map-support/register';
 import { App } from '@aws-cdk/core';
-import { VpcStack } from '../lib/vpc-cdk-stack';
-import { EksCdkStack } from '../lib/eks-cdk-stack';
+import { VpcStack } from '../lib/VpcStack';
+import { EksStack } from '../lib/EksStack';
 
 const app = new App();
 const vpcStack = new VpcStack(app, 'VpcCdkStack');
-new EksCdkStack(app, 'EksCdkStack', { vpc: vpcStack.vpc });
+new EksStack(app, 'EksCdkStack', { vpc: vpcStack.vpc });
